test: move agent cli detection out of tsx
