+++ src/tui/screens/agent-cli-detect.ts
@@ -1,0 +1,15 @@
+import { resolveAcpLaunch } from "../../core/acp-launch.js";
+
+/** Check if a CLI tool is installed or available through a bundled ACP adapter. */
+export function detectCli(name: string): boolean {
+  if (Bun.which(name) !== null) return true;
+  if (name === "claude" || name === "codex") {
+    try {
+      resolveAcpLaunch(name);
+      return true;
+    } catch {
+      return false;
+    }
+  }
+  return false;
+}
